LayoutTests/platform/wpe/imported/w3c/web-platform-tests/encrypted-media/clearkey-mp4-setmediakeys-to-multiple-video-elements.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: undefined is not an object (evaluating 'contentitem.audio')

Harness Error (FAIL), message = Unhandled rejection: undefined is not an object (evaluating 'contentitem.audio')
